:hammer: Rework and split to router props component
